rename: 패키지명 변경 - CarRacer.java -> Racer.java - package: game -> racer
